change weight datatype to double

+++ app/src/main/java/com/example/fitnesstracker/model/UserInformation.java
@@ -10,11 +10,11 @@ public class UserInformation {
     private int userId;
     private Date date;
     private int height;
-    private int weight;
+    private double weight;
     private int kfa; // Körperfettanteil in Prozent
 
     // Konstruktoren
-    public UserInformation(int id, int userId, Date date, int height, int weight, int kfa) {
+    public UserInformation(int id, int userId, Date date, int height, double weight, int kfa) {
         this.id = id;
         this.userId = userId;
         this.date = date;
@@ -23,7 +23,7 @@ public class UserInformation {
         this.kfa = kfa;
     }
 
-    public UserInformation(int id, int userId, String dateString, int height, int weight, int kfa) {
+    public UserInformation(int id, int userId, String dateString, int height, double weight, int kfa) {
         this.id = id;
         this.userId = userId;
         this.date = parseDate(dateString);  // Datum konvertieren
@@ -47,6 +47,6 @@ public class UserInformation {
     public int getUserId() { return userId; }
     public Date getDate() { return date; }
     public int getHeight() { return height; }
-    public int getWeight() { return weight; }
+    public double getWeight() { return weight; }
     public int getKfa() { return kfa; }
 }
